Rename and Document Gem
Rename project to "Chewie".

README
* Setup
* Filtering by Associations
* Format
* Combine

Yardoc
* All currently supported queries
