+++ lib/chewie/version.rb
@@ -1,3 +1,3 @@
-module EsBuilder
+module Chewie
   VERSION = "0.2.0"
 end
